支持从excel导入ConfigurationItem及相关重构，添加相关mocha测试用例
1,支持从excel导入ConfigurationItem及相关重构，添加相关mocha测试用例(6h)
2,支持processFlows的patch(2h)
3,支持ITService和ITServiceGroup的备份恢复(4h)
共12h

+++ script/import.sh
@@ -1,15 +1,53 @@
 #! /bin/bash
 
-#$NEO4J_HOME/bin/neo4j-shell -file ./cypher/dropSchema.cyp
+usage() { echo "Usage: $0 [--service_only ] --dir=20170123000348" 1>&2; exit 1; }
 
-curl -X DELETE --header "Content-Type: application/json" --url "http://localhost:3001/api/items"
+service_only=0
 
-cat ./data/$1/neo4j.dump | $NEO4J_HOME/bin/neo4j-shell
+while [ "$1" != "" ]; do
+    PARAM=`echo $1 | awk -F= '{print $1}'`
+    VALUE=`echo $1 | awk -F= '{print $2}'`
+    case $PARAM in
+        -h | --help)
+            usage
+            exit
+            ;;
+        --service_only)
+            service_only=1
+            ;;
+        --dir)
+            dir=$VALUE
+            ;;
+        *)
+            echo "ERROR: unknown parameter \"$PARAM\""
+            usage
+            exit 1
+            ;;
+    esac
+    shift
+done
 
-elasticdump \
-  --input=./data/$1/es.dump \
-  --output=http://localhost:9200/cmdb \
-  --type=data
+if [ -z "$dir" ]; then
+        usage
+        exit
+fi
+
+if [ $service_only -eq 0 ]; then
+    #$NEO4J_HOME/bin/neo4j-shell -file ./cypher/dropSchema.cyp
+    curl -X DELETE --header "Content-Type: application/json" --url "http://localhost:3001/api/items"
+    cat ./data/$dir/neo4j.dump | $NEO4J_HOME/bin/neo4j-shell
+    elasticdump \
+      --input=./data/$dir/es.dump \
+      --output=http://localhost:9200/cmdb \
+      --type=data
+    if [ $? -eq 0 ]; then
+        echo "importing data successfully!"
+    fi
+else
+   node ./importItService.js $dir
+fi
+
+
